jwt vehiculo done

+++ vehiculo/delete.php
@@ -1,31 +1,71 @@
 <?php
+
+$time1 = microtime(true);
+
 header("Access-Control-Allow-Origin: *");
 header("Content-Type: application/json; charset=UTF-8");
 header("Access-Control-Allow-Methods: POST");
 header("Access-Control-Max-Age: 3600");
 header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
  
+include_once '../config/core.php';
+include_once '../php-jwt/src/BeforeValidException.php';
+include_once '../php-jwt/src/ExpiredException.php';
+include_once '../php-jwt/src/SignatureInvalidException.php';
+include_once '../php-jwt/src/JWT.php';
+use \Firebase\JWT\JWT;
+
 include_once '../config/database.php';
+include_once '../objects/auditoria.php';
 include_once '../objects/vehiculo.php';
  
 $database = new Database();
 $db = $database->getConnection();
+$dbauditoria = $database->getConnection();
  
 $vehiculo = new Vehiculo($db);
+$auditoria= new Auditoria($dbauditoria);
  
 $data = json_decode(file_get_contents("php://input"));
  
 $vehiculo->vehiculo_id = $data->vehiculo_id;
+$jwt=isset($data->jwt) ? $data->jwt : "";
  
-if($vehiculo->delete()){
- 
-    http_response_code(200);
-    echo json_encode(array("message" => "vehiculo eliminado."));
-}
- 
-else{
- 
-    http_response_code(503);
-    echo json_encode(array("message" => "Error al eliminar el vehiculo."));
+if($jwt){
+
+    try {
+        $decoded = JWT::decode($jwt, $key, array('HS256'));
+        //obtener datos del jwt
+        $aux= get_object_vars($decoded);
+        $usu= get_object_vars($aux[data]);        
+        $usuario= $usu['nombre']." ".$usu['apellido'];
+
+        if($vehiculo->delete()){
+        
+            $time2= round(((microtime(true) - $time1)*1000), 2);
+            $auditoria->response_time= $time2;            
+            $auditoria->usuario= $usuario;
+            $auditoria->created = date('Y-m-d H:i:s');
+            //endpoint hay que cambiarlo para cada funcion seria el url de la pagina
+            $auditoria->endpoint= "localhost/prog1final/vehiculo/delete.php";
+            //agregar auditoria
+            $auditoria->create();
+
+            http_response_code(200);
+            echo json_encode(array("message" => "vehiculo eliminado."));
+        }
+        
+        else{
+        
+            http_response_code(503);
+            echo json_encode(array("message" => "Error al eliminar el vehiculo."));
+        }
+    }catch (Exception $e){
+        http_response_code(401);
+        echo json_encode(array(
+            "message" => "Token invalido.",
+            "error" => $e->getMessage()
+            ));
+    }
 }
 ?>
